Revert "Merge remote-tracking branch 'origin/master'"
This reverts commit 2beb3cc4e0051c3fb16b6006297b191290b47cef.

+++ app/frontend/models/Gitscan.php
@@ -13,48 +13,52 @@ class Gitscan extends ActiveRecord
 
     public static function gitscan($input)
     {
-        //$taskid = $input["taskid"];
 
-        $taskid = 3384;
+        global $a;
+        $a = "";
+        while ($a != "Done") {
+            $countdocker = "pgrep -c docker";
 
-        $task = Tasks::find()
-            ->where(['taskid' => $taskid])
-            ->limit(1)
-            ->one();
+            exec($countdocker, $count);
 
-        $token = "b54732426e4a8ad9ee3aa10baf51f2d8abb6a8cb"; //change later to pick up from DB
+            if ($count[0] < 5) {
 
+                system("systemctl start docker");
 
-        $task = json_decode($task["amass"], true);
+                $url = $input["url"];
+                $taskid = $input["taskid"];
 
-        foreach ($task as $json) {
-            $addresses[] = $json["name"];
+                $randomid = rand(1, 100000);
+
+                $gitscan = Tasks::find()
+                    ->where(['taskid' => $taskid])
+                    ->limit(1)
+                    ->one();
+
+                $token ="b54732426e4a8ad9ee3aa10baf51f2d8abb6a8cb"; //change later to pick up from DB
+
+                system('sudo /usr/bin/docker run --name=' . $randomid . ' abhartiya/tools_gitallsecrets -mergeOutput -token=' .
+                    $token . ' -repoURL=' . $url . ' >/dev/null && sudo /usr/bin/docker cp ' .
+                    $randomid . ':/root/results.txt /root/gitscan' . $randomid . ' && sudo /usr/bin/docker rm ' .
+                    $randomid . '&& sudo mv /root/gitscan' . $randomid . ' /var/www/output/docker/' . $randomid . ' ');
+
+                $output = file_get_contents("/var/www/output/docker/" . $randomid);
+
+                $date_end = date("Y-m-d H-i-s");
+
+                $gitscan->gitscan_status = "Done.";
+                $gitscan->gitscan = $output;
+                $gitscan->date = $date_end;
+
+                system("find /var/www/output/docker/ -name " . $randomid . " -delete &");
+                $a = "Done";
+                return $gitscan->save();
+
+            } else sleep(25);
         }
-        var_dump($addresses);
-        return 1;
-
-
-
-
-
-
-        /*system('sudo /usr/bin/docker run --name=' . $randomid . ' abhartiya/tools_gitallsecrets -mergeOutput -token=' .
-               $token . ' -repoURL=' . $url . ' >/dev/null && sudo /usr/bin/docker cp ' .
-            $randomid . ':/root/results.txt /root/gitscan' . $randomid . ' && sudo /usr/bin/docker rm ' .
-            $randomid . '&& sudo mv /root/gitscan' . $randomid . ' /var/www/output/docker/' . $randomid . ' ');
-
-        $output = file_get_contents("/var/www/output/docker/" . $randomid);
-*/
-
-        $task->gitscan_status = "Done.";
-        //$gitscan->gitscan = $output;
-        $task->date = date("Y-m-d H-i-s");
-
-                //system("find /var/www/output/docker/ -name " . $randomid . " -delete &");
-        $task->save();
-
-        return 1;
     }
+
+
 }
